Complete rewrite
The project has been changed from being a server side yesod application to a
client side application. GHCJS is now used to compile the code to JS. Macro.hs
has almost completely been rewritten, with an emphasis to simplify the code,
which has been done through the use of fclabels and better use of monads. The UI
has been improved and now has a drag and drop interface, and is more similar to
the tables as they appeared in Turing's paper.